Update README.md
Figuring out what commits are and how they are used. Looks like you can commit changes to a branch and then later commit the changes in the branch to the main -- merging everything. That gives you a chance to avoid mistakes like deleting stuff and then having to rebuild it.